split out start
